+++ src/collective/cover/testing.py
@@ -35,6 +35,13 @@ else:
     DEXTERITY_ONLY = os.environ.get('DEXTERITY_ONLY') is not None
 
 try:
+    pkg_resources.get_distribution('plone.app.widgets')
+except pkg_resources.DistributionNotFound:
+    HAS_WIDGETS = False
+else:
+    HAS_WIDGETS = True
+
+try:
     pkg_resources.get_distribution('Products.PloneFormGen')
 except pkg_resources.DistributionNotFound:
     HAS_PFG = False
@@ -133,6 +140,12 @@ class Fixture(PloneSandboxLayer):
                 self.loadZCML(package=plone.app.contenttypes)
                 z2.installProduct(app, 'Products.DateRecurringIndex')
 
+            if HAS_WIDGETS:
+                import plone.app.widgets
+                self.loadZCML(package=plone.app.widgets)
+                import plone.formwidget.querystring
+                self.loadZCML(package=plone.formwidget.querystring)
+
             if HAS_PFG:
                 import Products.PloneFormGen
                 self.loadZCML(package=Products.PloneFormGen)
@@ -154,6 +167,10 @@ class Fixture(PloneSandboxLayer):
         else:
             if DEXTERITY_ONLY:
                 self.applyProfile(portal, 'plone.app.contenttypes:default')
+
+            if HAS_WIDGETS:
+                self.applyProfile(portal, 'plone.app.widgets:default')
+                self.applyProfile(portal, 'plone.formwidget.querystring:default')  # noqa
 
             if HAS_PFG:
                 self.applyProfile(portal, 'Products.PloneFormGen:default')